Fixed the error uninitialized member: 'ctrlKey' when mobile safari triggers the keydown event with non-standard CustomEvent type argument; Added GetDefault method to suEl to make handling nonexistent members easier
